feat: made TradingTable display all stocks within the Exchange
