+++ src/main/java/org/edu/ntnu/idatt2003/group49/millions/view/components/MillionsTable/OwnedSharesTable.java
@@ -3,11 +3,16 @@ package org.edu.ntnu.idatt2003.group49.millions.view.components.MillionsTable;
 import javafx.scene.control.*;
 import org.edu.ntnu.idatt2003.group49.millions.controller.Navigator;
 import org.edu.ntnu.idatt2003.group49.millions.model.exchange.Share;
+import org.edu.ntnu.idatt2003.group49.millions.view.components.MillionsTable.factory.OwnedSharesColumnFactory;
+import org.edu.ntnu.idatt2003.group49.millions.view.components.MillionsTable.factory.TableColumnFactory;
 
 import java.util.List;
 
 public class OwnedSharesTable extends MillionsTable<Share> {
-  public OwnedSharesTable(Navigator navigator) {
+  private final OwnedSharesColumnFactory columnFactory;
+
+  public OwnedSharesTable(Navigator navigator, OwnedSharesColumnFactory columnFactory) {
+    this.columnFactory = columnFactory;
     super(navigator);
   }
